move custom manager types to /custom
distribute types CustomManager and RegexManager
