+++ lib/modules/manager/custom/regex/index.ts
@@ -1,5 +1,4 @@
 import is from '@sindresorhus/is';
-import type { RegexManagerTemplates } from '../../../../config/types';
 import type {
   ExtractConfig,
   PackageDependency,
@@ -7,7 +6,7 @@ import type {
   Result,
 } from '../../types';
 import { handleAny, handleCombination, handleRecursive } from './strategies';
-import type { RegexManagerConfig } from './types';
+import type { RegexManagerConfig, RegexManagerTemplates } from './types';
 import { validMatchFields } from './utils';
 
 export const defaultConfig = {
