+++ lib/modules/manager/custom/regex/strategies.ts
@@ -84,7 +84,7 @@ function processRecursive(parameters: RecursionParameter): PackageDependency[] {
     regexes,
     config,
   }: RecursionParameter = parameters;
-  // abort if we have no matchString anymore
+  // abort if we have no matched strings anymore
   if (regexes.length === index) {
     const result = createDependency(
       {
